requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2
ipython>=8.18

# Code quality
black>=23.12
isort>=5.13
flake8>=6.1
mypy>=1.7

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3
faker>=22.0

# Django stubs for mypy
django-stubs>=4.2
